+++ src/simplecep/__init__.py
@@ -23,8 +23,8 @@ DEFAULT_TIMEOUT = 2.0
 DEFAULT_PROVIDERS = (
     # try to resolve the CEP using this one first
     {correios_sigep_cep_provider},
-    # if the above provider fails, try using these both in parallel using threads
-    # the first returned result will be used
+    # if the above provider fails, use third-party ones
+    # in parallel using threads. the first received result will be used
     {republicavirtual_cep_provider, viacep_cep_provider},
 )